[REFACTOR] User 관련 패키지 그룹화
-TripStyle를 User 패키지로 병합
