+++ src/main/java/com/triprecord/triprecord/record/entity/RecordComment.java
@@ -2,7 +2,7 @@ package com.triprecord.triprecord.record.entity;
 
 
 import com.triprecord.triprecord.global.util.EntityBaseTime;
-import com.triprecord.triprecord.user.User;
+import com.triprecord.triprecord.user.entity.User;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
