+++ src/main/java/com/triprecord/triprecord/schedule/entity/ScheduleComment.java
@@ -1,8 +1,7 @@
 package com.triprecord.triprecord.schedule.entity;
 
 import com.triprecord.triprecord.global.util.EntityBaseTime;
-import com.triprecord.triprecord.schedule.entity.Schedule;
-import com.triprecord.triprecord.user.User;
+import com.triprecord.triprecord.user.entity.User;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
