+++ src/main/java/com/triprecord/triprecord/user/UserService.java
@@ -3,7 +3,9 @@ package com.triprecord.triprecord.user;
 import com.triprecord.triprecord.basicproflie.BasicProfileRepository;
 import com.triprecord.triprecord.global.config.jwt.JwtProvider;
 import com.triprecord.triprecord.global.config.jwt.UserAuthentication;
-import com.triprecord.triprecord.tripstyle.TripStyleRepository;
+import com.triprecord.triprecord.user.repository.TripStyleRepository;
+import com.triprecord.triprecord.user.entity.User;
+import com.triprecord.triprecord.user.repository.UserRepository;
 import jakarta.security.auth.message.AuthException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
