+++ src/main/java/com/triprecord/triprecord/user/repository/TripStyleRepository.java
@@ -1,5 +1,6 @@
-package com.triprecord.triprecord.tripstyle;
+package com.triprecord.triprecord.user.repository;
 
+import com.triprecord.triprecord.user.entity.TripStyle;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TripStyleRepository extends JpaRepository<TripStyle, Long> {
